Implementation_1/requirements.txt
torch==2.0.1
torchvision==0.15.2
Pillow==9.4.0
numpy==1.24.4
scikit-learn==1.2.0
tqdm==4.65.0
